Use parseFloat

+++ functions/halfcounter.js
@@ -1,7 +1,7 @@
 export async function onRequest(context) {
     const { MYKVSTORE } = context.env;
     let counterValue = await MYKVSTORE.get('counter');
-    counterValue = parseInt(counterValue) || 0;
+    counterValue = parseFloat(counterValue) || 0;
     counterValue = counterValue/2;
     await MYKVSTORE.put('counter', counterValue);
     return new Response(counterValue);
